requirements.txt
google-adk
python-dotenv
duckduckgo-search
